+++ core/blong-int-adapter/k8s/adapter/cluster.ts
@@ -1,0 +1,17 @@
+import {adapter} from '@feasibleone/blong';
+
+export default adapter<{
+    k8s: {
+        kubeconfig?: string;
+        namespace?: string;
+    };
+}>(api => ({
+    extends: 'adapter.k8s',
+    activation: {
+        default: {
+            k8s: {},
+            namespace: 'cluster',
+            imports: [],
+        },
+    },
+}));